% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops_get.R
\name{find_}
\alias{find_}
\title{Return First Operator with Name
Returns NULL if Operator not found}
\usage{
find_(.data, name)
}
\description{
Return First Operator with Name
Returns NULL if Operator not found
}
